refactor: move `CheckForUpdatesHelper` out to main package
Also renamed it to `UpdateChecker` to better reflect its purpose.
