+++ src/main/java/site/overwrite/auditranscribe/MainApplication.java
@@ -25,7 +25,6 @@ import site.overwrite.auditranscribe.io.IOMethods;
 import site.overwrite.auditranscribe.io.PropertyFile;
 import site.overwrite.auditranscribe.io.data_files.DataFiles;
 import site.overwrite.auditranscribe.misc.MyLogger;
-import site.overwrite.auditranscribe.views.main.CheckForUpdatesHelper;
 import site.overwrite.auditranscribe.views.main.scene_switching.SceneSwitcher;
 import site.overwrite.auditranscribe.views.setup_wizard.SetupWizard;
 
@@ -66,7 +65,7 @@ public class MainApplication extends Application {
         }
 
         // Check if there are any updates
-        CheckForUpdatesHelper.checkForUpdates(currentVersion);
+        UpdateChecker.checkForUpdates(currentVersion);
 
         // Start scene handler
         SceneSwitcher sceneHandler = new SceneSwitcher(currentVersion);
